+++ lib/NaiveBayesTextTagger.jsm
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 "use strict";
 
-const {toksToTfIdfVector} = ChromeUtils.import("resource://activity-stream/lib/TfIdfVectorizer.jsm", {});
+const {toksToTfIdfVector} = ChromeUtils.import("resource://activity-stream/lib/Tokenize.jsm", {});
 
 this.NaiveBayesTextTagger = class NaiveBayesTextTagger {
   constructor(model) {
