+++ test/unit/lib/NmfTextTagger.test.js
@@ -1,5 +1,5 @@
 import {NmfTextTagger} from "lib/NmfTextTagger.jsm";
-import {tokenize} from "lib/TfIdfVectorizer.jsm";
+import {tokenize} from "lib/Tokenize.jsm";
 
 const EPSILON = 0.00001;
